Module google_acceleratedmobilepageurl1::api

Structs

Acceleratedmobilepageurl
Central instance to access all Acceleratedmobilepageurl related resource activities

AmpUrl
AMP URL response for a requested URL.

AmpUrlBatchGetCall
Returns AMP URL(s) and equivalent AMP Cache URL(s).

AmpUrlError
AMP URL Error resource for a requested URL that couldn't be found.

AmpUrlMethods
A builder providing access to all methods supported on ampUrl resources. It is not used directly, but through the Acceleratedmobilepageurl hub.

BatchGetAmpUrlsRequest
AMP URL request for a batch of URLs.

BatchGetAmpUrlsResponse
Batch AMP URL response.